requirements.txt
cachetools==5.2.0
google-auth==2.6.6
google-auth-oauthlib==0.5.1
gspread==5.4.0
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
requests-oauthlib==1.3.1
rsa==4.8
termcolor==1.1.0
